scripts: Allow packages to have their own condarc
It can be a single condarc for all platforms: $PACKAGE/condarc, but if
$PACKAGE/condarc_$TRAVIS_OS_NAME exists, it will be used instead (with
TRAVIS_OS_NAME being 'linux', 'osx' or 'windows').

All of the condarc files placed in proper directories influence conda.
Their importance and conflict resolving scheme is based on where they're
placed. For details, see: "Using the .condarc conda configuration file".

Changes made with 'conda config --add/--set' are saved in the ~/.condarc
file, which has the highest priority if there's no environment
(conda-build ignores CONDARC environment variable). That's why it has to
be moved to a path which has lower priority and package's condarc is
used as the most important configuration file.