Add Renode package
The package uses a script to download the current Renode recipe before
executing 'conda build' and patch it for building with Travis (there are
mostly minor differences e.g. because of the 'msbuild' version).

The most important difference is, the Renode on Linux and macOS can't
be built with GUI support. That's because GtkSharp libraries can't be
provided through Conda. Such headless version of Renode can still be
used using telnet as a terminal, however.

Prescript doesn't influence the build itself. All files used to create
the recipe from the one found in Renode repository (including the
prescript itself) are actually cleaned before executing 'conda build',
so they won't be packaged either.